refactor: Extract 48 helper functions from cli.py to cli_helpers.py (Phase 2)
Reduces cli.py from 4,974 LOC to 2,500 LOC (49.7% reduction).

Extracted 48 helper functions organized by category:
- VM operations (9)
- Config/validation (12)
- Display/UI (4)
- Sync operations (5)
- Deletion operations (5)
- Clone operations (5)
- NLP/AI operations (2)
- SSH/Keys (3)
- Batch operations (2)
- Other (2)

Changes:
- cli.py: 4,974 → 2,500 LOC (-2,474 LOC, -49.7%)
- cli_helpers.py: 282 → 2,751 LOC (+2,469 LOC)
- Added 48 imports from cli_helpers to cli.py
- Fixed all missing imports and cleaned up unused imports
- All pre-commit hooks pass

Progress to target:
- Target: < 1,500 LOC
- Current: 2,500 LOC
- Remaining: 1,000 LOC to remove (40% more reduction needed)

Next: Phase 3 - move Click command implementations to commands/ modules
